+++ src/Rules.php
@@ -10,4 +10,6 @@ class Rules
     const NUMBER = 'Validator\Rules\Number';
     const BOOLEAN = 'Validator\Rules\Boolean';
     const EMAIL = 'Validator\Rules\Email';
+    const MAX = 'Validator\Rules\Max';
+    const MIN = 'Validator\Rules\Min';
 }