+++ src/Validate.php
@@ -19,9 +19,16 @@ class Validate
      */
     protected $messages = [];
 
+    protected $parser;
+
+    public function __construct(Parser $parser)
+    {
+        $this->parser = $parser;
+    }
+
     /**
      * wrapping the validation.
-     * 
+     *
      * @param array $data
      * @param array $rules
      */
@@ -55,10 +62,10 @@ class Validate
 
     /**
      * Determine if the data is passed.
-     * 
+     *
      * @return bool
      */
-    public function isPassed() : bool
+    public function isPassed(): bool
     {
         if ($this->messages) {
             return false;
@@ -69,39 +76,45 @@ class Validate
 
     /**
      * Get the error messages.
-     * 
+     *
      * @return array $messages
      */
-    public function getMessages() : array
+    public function getMessages(): array
     {
         return $this->messages;
     }
 
     /**
      * Fitching the rules for the field.
-     * 
+     *
      * @param array $rules
-     * @param mix $rules
-     * 
+     * @param mix   $rules
+     *
      * @return void
      */
-    private function ruleFitcher(array $rules, $attribute) : void
+    private function ruleFitcher(array $rules, $attribute): void
     {
         foreach ($rules as $rule) {
-            $this->ruleChecker(new $rule(), $attribute, $this->data[$attribute]);
+            if ($this->parser->has($rule, '|')) {
+                $arguments = $this->parser->getArguments($rule);
+                $rule = $this->parser->getRule($rule);
+                $this->ruleChecker(new $rule(...$arguments), $attribute, $this->data[$attribute]);
+            } else {
+                $this->ruleChecker(new $rule(), $attribute, $this->data[$attribute]);
+            }
         }
     }
 
     /**
      * Fill the message if the rule not valid.
-     * 
+     *
      * @param Validator\RuleInterface $rule
-     * @param mix $attribute
-     * @param mix $value
-     * 
+     * @param mix                     $attribute
+     * @param mix                     $value
+     *
      * @return void
      */
-    private function ruleChecker(RuleInterface $rule, $attribute, $value) : void
+    private function ruleChecker(RuleInterface $rule, $attribute, $value): void
     {
         if (!$rule->isValid($value)) {
             $this->messages[$attribute][] = $rule->getMessage($attribute);