+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Validator - Change Log
 
 - Version [0.0.0](#what's-new-in-validator-000-release-date-jan-27-2020) (release date: Jan 27, 2020)
+- Version [1.0.0](#what's-new-in-validator-100-release-date-jan-27-2020) (release date: Jan 27, 2020)
 
 
 ## What's new in Validator v0.0.0 _(release date: Jan 27, 2020)_
 it's beta version for validation, it follow the Strategy Pattern for changing the rule in the run time. and it dose not support any dynamic argument for rule.
+
+## What's new in Validator v1.0.0 _(release date: Jan 27, 2020)_
+The rules can take dynamic argument like max value, and Facades Pattern for deal with complex object.
+### New Rules
+- Max (take argument for max value).
+- Min (take argument for min value).